Added ability to store tokens for the last logged in user after logging out

+++ Softeq.XToolkit.iOS.DefaultAuthorization/SecuredTokenManager.cs
@@ -10,21 +10,32 @@ namespace Softeq.iOS.DefaultAuthorization
     {
         private const string SESSION_TOKEN_KEY = "SessionToken";
         private const string REFRESH_TOKEN_KEY = "RefreshToken";
+        private const string TOKENS_MARKED_FOR_DELETION_KEY = "TokensMarkedForDeletion";
 
         public string Token { get; private set; }
         public string RefreshToken { get; private set; }
+        public bool AreTokensMarkedForDeletion { get; private set; }
 
         public SecuredTokenManager()
         {
             RestoreTokens();
         }
 
-        public Task ResetTokensAsync()
+        public Task ResetTokensAsync(bool shouldDeleteTokensPermanently = true)
         {
-            UpdateTokens(null, null);
+            if (shouldDeleteTokensPermanently)
+            {
+                UpdateTokens(null, null, false);
 
-            CrossSecureStorage.Current.DeleteKey(SESSION_TOKEN_KEY);
-            CrossSecureStorage.Current.DeleteKey(REFRESH_TOKEN_KEY);
+                CrossSecureStorage.Current.DeleteKey(SESSION_TOKEN_KEY);
+                CrossSecureStorage.Current.DeleteKey(REFRESH_TOKEN_KEY);
+                CrossSecureStorage.Current.DeleteKey(TOKENS_MARKED_FOR_DELETION_KEY);
+            }
+            else
+            {
+                AreTokensMarkedForDeletion = true;
+                CrossSecureStorage.Current.SetValue(TOKENS_MARKED_FOR_DELETION_KEY, "true");
+            }
 
             return Task.CompletedTask;
         }
@@ -40,7 +51,7 @@ namespace Softeq.iOS.DefaultAuthorization
                     "Please check iOS settings by the following link: https://github.com/sameerkapps/SecureStorage/issues/31#issuecomment-366205742");
             }
 
-            UpdateTokens(token, refreshToken);
+            UpdateTokens(token, refreshToken, AreTokensMarkedForDeletion);
 
             return Task.CompletedTask;
         }
@@ -48,13 +59,15 @@ namespace Softeq.iOS.DefaultAuthorization
         private void RestoreTokens()
         {
             UpdateTokens(CrossSecureStorage.Current.GetValue(SESSION_TOKEN_KEY),
-                CrossSecureStorage.Current.GetValue(REFRESH_TOKEN_KEY));
+                         CrossSecureStorage.Current.GetValue(REFRESH_TOKEN_KEY),
+                         CrossSecureStorage.Current.HasKey(TOKENS_MARKED_FOR_DELETION_KEY));
         }
 
-        private void UpdateTokens(string token, string refreshToken)
+        private void UpdateTokens(string token, string refreshToken, bool areTokensMarkedForDeletion)
         {
             Token = token;
             RefreshToken = refreshToken;
+            AreTokensMarkedForDeletion = areTokensMarkedForDeletion;
         }
     }
 }
